// src/scripts/enums/Positions.ts
/*
 * Copyright (c) 2021 EGT Digital - All rights reserved.
 * 
 * File: Positions.ts
 * 
 * Written by Deyan Ivanov <deyan.ivanov@egt-digital.com> on  28.01.24 3:44 pm.
 * Last modified: 28.01.24 3:44 pm.
 * 
 * More information at: https://egt-digital.com/
 */
export enum EPositions {
    POS0 = '0',
    POS90 = '90',
    POS180 = '180',
    POS270 = '270',
}